Move annotations to engine
