+++ kotest-framework/kotest-framework-api/api/kotest-framework-api.api
@@ -22,20 +22,6 @@ public final class io/kotest/core/NamedTag : io/kotest/core/Tag {
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class io/kotest/core/Platform : java/lang/Enum {
-	public static final field JS Lio/kotest/core/Platform;
-	public static final field JVM Lio/kotest/core/Platform;
-	public static final field Native Lio/kotest/core/Platform;
-	public static final field WasmJs Lio/kotest/core/Platform;
-	public static fun getEntries ()Lkotlin/enums/EnumEntries;
-	public static fun valueOf (Ljava/lang/String;)Lio/kotest/core/Platform;
-	public static fun values ()[Lio/kotest/core/Platform;
-}
-
-public final class io/kotest/core/PlatformKt {
-	public static final fun getPlatform ()Lio/kotest/core/Platform;
-}
-
 public abstract class io/kotest/core/Tag {
 	public static final field Companion Lio/kotest/core/Tag$Companion;
 	public fun <init> ()V
@@ -829,96 +815,6 @@ public final class io/kotest/core/TupleKt {
 
 public final class io/kotest/core/WriteLogKt {
 	public static final fun writeLog (Lkotlin/time/TimeMark;Ljava/lang/Throwable;Lkotlin/jvm/functions/Function0;)V
-}
-
-public final class io/kotest/core/annotation/AlwaysEnabledCondition : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public abstract interface annotation class io/kotest/core/annotation/Description : java/lang/annotation/Annotation {
-	public abstract fun value ()Ljava/lang/String;
-}
-
-public abstract interface annotation class io/kotest/core/annotation/DisplayName : java/lang/annotation/Annotation {
-	public abstract fun name ()Ljava/lang/String;
-}
-
-public abstract interface annotation class io/kotest/core/annotation/DoNotParallelize : java/lang/annotation/Annotation {
-}
-
-public abstract interface class io/kotest/core/annotation/EnabledCondition {
-	public abstract fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public abstract interface annotation class io/kotest/core/annotation/EnabledIf : java/lang/annotation/Annotation {
-	public abstract fun enabledIf ()Ljava/lang/Class;
-}
-
-public abstract interface annotation class io/kotest/core/annotation/Ignored : java/lang/annotation/Annotation {
-	public abstract fun reason ()Ljava/lang/String;
-}
-
-public abstract interface annotation class io/kotest/core/annotation/Isolate : java/lang/annotation/Annotation {
-}
-
-public final class io/kotest/core/annotation/NeverEnabled : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public abstract interface annotation class io/kotest/core/annotation/RequiresPlatform : java/lang/annotation/Annotation {
-	public abstract fun values ()[Lio/kotest/core/Platform;
-}
-
-public abstract interface annotation class io/kotest/core/annotation/RequiresTag : java/lang/annotation/Annotation {
-	public abstract fun values ()[Ljava/lang/String;
-}
-
-public abstract interface annotation class io/kotest/core/annotation/Tags : java/lang/annotation/Annotation {
-	public abstract fun values ()[Ljava/lang/String;
-}
-
-public final class io/kotest/core/annotation/displayname/WrapperKt {
-	public static final fun getWrapper (Lio/kotest/core/annotation/DisplayName;)Ljava/lang/String;
-}
-
-public final class io/kotest/core/annotation/enabledif/LinuxCondition : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public final class io/kotest/core/annotation/enabledif/MacCondition : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public final class io/kotest/core/annotation/enabledif/NotLinuxCondition : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public final class io/kotest/core/annotation/enabledif/NotMacCondition : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public final class io/kotest/core/annotation/enabledif/NotWindowsCondition : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public final class io/kotest/core/annotation/enabledif/WindowsCondition : io/kotest/core/annotation/EnabledCondition {
-	public fun <init> ()V
-	public fun enabled (Lkotlin/reflect/KClass;)Z
-}
-
-public final class io/kotest/core/annotation/enabledif/WrapperKt {
-	public static final fun getWrapper (Lio/kotest/core/annotation/EnabledIf;)Lkotlin/reflect/KClass;
-}
-
-public final class io/kotest/core/annotation/requirestag/WrapperKt {
-	public static final fun getWrapper (Lio/kotest/core/annotation/RequiresTag;)[Ljava/lang/String;
 }
 
 public abstract interface class io/kotest/core/concurrency/CoroutineDispatcherFactory {
