--- kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/annotation/displayname/wrapper.kt
@@ -1,5 +1,0 @@
-package io.kotest.core.annotation.displayname
-
-import io.kotest.core.annotation.DisplayName
-
-expect val DisplayName.wrapper: String?
