--- kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/annotation/enabledif/Wrapper.kt
@@ -1,7 +1,0 @@
-package io.kotest.core.annotation.enabledif
-
-import io.kotest.core.annotation.EnabledCondition
-import io.kotest.core.annotation.EnabledIf
-import kotlin.reflect.KClass
-
-expect val EnabledIf.wrapper: KClass<out EnabledCondition>
