--- kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/annotation/requirestag/Wrapper.kt
@@ -1,5 +1,0 @@
-package io.kotest.core.annotation.requirestag
-
-import io.kotest.core.annotation.RequiresTag
-
-expect val RequiresTag.wrapper: Array<out String>
