--- kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/spec/annnotations.kt
@@ -1,7 +1,0 @@
-package io.kotest.core.spec
-
-// these are backwards compatible typealiases
-
-typealias DisplayName = io.kotest.core.annotation.DisplayName
-typealias DoNotParallelize = io.kotest.core.annotation.DoNotParallelize
-typealias Isolate = io.kotest.core.annotation.Isolate
