--- kotest-framework/kotest-framework-api/src/jvmMain/kotlin/io/kotest/core/annotation/displayname/wrapper.kt
@@ -1,6 +1,0 @@
-package io.kotest.core.annotation.displayname
-
-import io.kotest.core.annotation.DisplayName
-
-actual val DisplayName.wrapper: String?
-   get() = name
