--- kotest-framework/kotest-framework-api/src/jvmMain/kotlin/io/kotest/core/annotation/enabledif/Wrapper.kt
@@ -1,8 +1,0 @@
-package io.kotest.core.annotation.enabledif
-
-import io.kotest.core.annotation.EnabledCondition
-import io.kotest.core.annotation.EnabledIf
-import kotlin.reflect.KClass
-
-actual val EnabledIf.wrapper: KClass<out EnabledCondition>
-   get() = enabledIf
