--- kotest-framework/kotest-framework-api/src/jvmMain/kotlin/io/kotest/core/annotation/requirestag/wrapper.kt
@@ -1,6 +1,0 @@
-package io.kotest.core.annotation.requirestag
-
-import io.kotest.core.annotation.RequiresTag
-
-actual val RequiresTag.wrapper: Array<out String>
-   get() = values
