--- kotest-framework/kotest-framework-api/src/jvmTest/kotlin/io/kotest/core/test/EnabledSpec.kt
@@ -1,19 +1,0 @@
-package io.kotest.core.test
-
-import io.kotest.core.spec.style.FunSpec
-
-class EnabledSpec : FunSpec({
-
-//   test("multiple enabled should resolve to enabled") {
-//      val result = (0..3).map { Enabled.enabled }.let { Enabled.fold(it) }
-//      result.isEnabled shouldBe true
-//      result.reason.shouldBeEmpty()
-//   }
-//
-//   test("multiple enabled and a single disabled should resolve to disabled") {
-//      val reason = UUID.randomUUID().toString()
-//      val result = ((0..3).map { Enabled.enabled } + Enabled.disabled(reason)).let { Enabled.fold(it.shuffled()) }
-//      result.isEnabled shouldBe false
-//      result.reason shouldBe reason
-//   }
-})
