+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/core/annotation/Isolate.kt
