+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/spec/interceptor/ref/EnabledIfInterceptor.kt
@@ -1,7 +1,6 @@
 package io.kotest.engine.spec.interceptor.ref
 
 import io.kotest.core.annotation.EnabledIf
-import io.kotest.core.annotation.enabledif.wrapper
 import io.kotest.core.config.ExtensionRegistry
 import io.kotest.core.spec.SpecRef
 import io.kotest.core.test.TestCase
@@ -33,7 +32,7 @@ internal class EnabledIfInterceptor(
 
       val enabled = ref.kclass
          .annotation<EnabledIf>(IncludingAnnotations, IncludingSuperclasses)
-         ?.wrapper
+         ?.enabledIf
          ?.newInstanceNoArgConstructor()
          ?.enabled(ref.kclass) ?: true
