+++ kotest-framework/kotest-framework-engine/src/jsMain/kotlin/io/kotest/core/platform.js.kt
