+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/io/kotest/engine/DerivedTagNameTest.kt
@@ -1,4 +1,4 @@
-package com.sksamuel.kotest.core.api
+package io.kotest.engine
 
 import io.kotest.core.Tag
 import io.kotest.core.spec.style.FunSpec
