+++ kotest-framework/kotest-framework-engine/src/wasmJsMain/kotlin/io/kotest/core/platform.wasmJs.kt
@@ -1,0 +1,3 @@
+package io.kotest.core
+
+actual val platform: Platform = Platform.WasmJs
